+++ app/code/Magento/Multishipping/Block/Checkout/Billing.php
@@ -131,7 +131,7 @@ class Billing extends \Magento\Payment\Block\Form\Container
      */
     public function getQuoteBaseGrandTotal()
     {
-        return (double)$this->getQuote()->getBaseGrandTotal();
+        return (float)$this->getQuote()->getBaseGrandTotal();
     }
 
     /**
