+++ app/code/Magento/Paypal/Model/Cart.php
@@ -134,7 +134,7 @@ class Cart extends \Magento\Payment\Model\Cart
             // aggregate item price if item qty * price does not match row total
             $itemBaseRowTotal = $item->getOriginalItem()->getBaseRowTotal();
             if ($amount * $qty != $itemBaseRowTotal) {
-                $amount = (double)$itemBaseRowTotal;
+                $amount = (float)$itemBaseRowTotal;
                 $subAggregatedLabel = ' x' . $qty;
                 $qty = 1;
             }
@@ -178,8 +178,8 @@ class Cart extends \Magento\Payment\Model\Cart
         \Magento\Payment\Model\Cart\SalesModel\SalesModelInterface $salesEntity
     ) {
         $dataContainer = $salesEntity->getTaxContainer();
-        $this->addTax((double)$dataContainer->getBaseDiscountTaxCompensationAmount());
-        $this->addTax((double)$dataContainer->getBaseShippingDiscountTaxCompensationAmount());
+        $this->addTax((float)$dataContainer->getBaseDiscountTaxCompensationAmount());
+        $this->addTax((float)$dataContainer->getBaseShippingDiscountTaxCompensationAmount());
     }
 
     /**
