+++ app/code/Magento/Paypal/Model/Express/Checkout.php
@@ -983,7 +983,7 @@ class Checkout
 
         foreach ($address->getGroupedAllShippingRates() as $group) {
             foreach ($group as $rate) {
-                $amount = (double)$rate->getPrice();
+                $amount = (float)$rate->getPrice();
                 if ($rate->getErrorMessage()) {
                     continue;
                 }
