+++ app/code/Magento/Sales/Model/Order/CreditmemoFactory.php
@@ -199,8 +199,8 @@ class CreditmemoFactory
     protected function initData($creditmemo, $data)
     {
         if (isset($data['shipping_amount'])) {
-            $creditmemo->setBaseShippingAmount((double)$data['shipping_amount']);
-            $creditmemo->setBaseShippingInclTax((double)$data['shipping_amount']);
+            $creditmemo->setBaseShippingAmount((float)$data['shipping_amount']);
+            $creditmemo->setBaseShippingInclTax((float)$data['shipping_amount']);
         }
         if (isset($data['adjustment_positive'])) {
             $creditmemo->setAdjustmentPositive($data['adjustment_positive']);
